将所有float全部改为double

+++ main.cpp
@@ -22,7 +22,7 @@ const int ans_task[6][2] = {{3, 6},
 
 
 //计算距离
-inline float distance(double x1, double y1, double x2, double y2) {
+inline double distance(double x1, double y1, double x2, double y2) {
     return pow(pow(x1 - x2, 2) + pow(y1 - y2, 2), 0.5);
 }
 
@@ -35,8 +35,8 @@ public:
 
     int init_stage_id;//获取物品工作台类型
     int target_stage_id;//出售物品工作台类型
-    float target_x;//目标工作台x坐标
-    float target_y;//目标工作台y坐标
+    double target_x;//目标工作台x坐标
+    double target_y;//目标工作台y坐标
 };
 
 //任务类构造函数
@@ -75,8 +75,8 @@ Task Producer::get_task() {
 class Stage {
 public:
     int stage_id;//工作台类型
-    float pos_x;//x坐标
-    float pos_y;//y坐标
+    double pos_x;//x坐标
+    double pos_y;//y坐标
     constexpr static int not_producing = -1;
     constexpr static int blocking = 0;
     int rest_time;//剩余生产时间, -1表示不在生产, 0表示生产因输出格满而阻塞, >=0表示剩余生产帧数
@@ -156,14 +156,14 @@ class Robot {
 public:
     int stage_id;//所处工作台, -1 表示当前没有处于任何工作台附近, [0,工作台总数-1] 表示某工作台的下标。当前机器人的所有购买、出售行为均针对该工作台进行。
     int object_id;//物品id, 0 表示未携带物品, 1-7 表示对应物品
-    float time_value_coef;//时间价值系数
-    float crash_value_coef;//碰撞价值系数
-    float v_rad;//角速度, 逆时针为正, 单位弧度每秒, [-pi, pi]
-    float v_x;//线速度
-    float v_y;//线速度
-    float pos_rad;//朝向
-    float pos_x;//x坐标
-    float pos_y;//y坐标
+    double time_value_coef;//时间价值系数
+    double crash_value_coef;//碰撞价值系数
+    double v_rad;//角速度, 逆时针为正, 单位弧度每秒, [-pi, pi]
+    double v_x;//线速度
+    double v_y;//线速度
+    double pos_rad;//朝向
+    double pos_x;//x坐标
+    double pos_y;//y坐标
     bool is_busy;//空闲状态
     int id; // 机器人id[0, 3], 目前一共只有4个机器人
     Stage *target_stage;
@@ -177,11 +177,11 @@ public:
     Robot() : id(-1), pos_x(-1), pos_y(-1), stage_id(no_stage), object_id(no_object) {
     }
 
-    void print_forward(float v) {
+    void print_forward(double v) {
         cout << "forward" << ' ' << id << ' ' << v << endl;
     }
 
-    void print_rotate(float v) {
+    void print_rotate(double v) {
         cout << "rotate" << ' ' << id << ' ' << v << endl;
     }
 
@@ -292,8 +292,8 @@ bool material_exist(Stage &stage, Robot &robot) {
 //找到最近的任务目标点
 void find_nearest_pos(Map &map, Robot &robot, bool flag) {
     int target_stage_id;
-    float dis;
-    float min_distance = INT32_MAX;
+    double dis;
+    double min_distance = INT32_MAX;
     //为真是获取物品位置，为假是出售位置
     if (flag) {
         target_stage_id = robot.task.init_stage_id;
@@ -321,7 +321,7 @@ void find_nearest_pos(Map &map, Robot &robot, bool flag) {
 
 
 // 子函数：处理char型数组，按照空格切割并翻译为浮点数
-void parse_char(char *line, float *temp_arr) {
+void parse_char(char *line, double *temp_arr) {
     char delims[] = " ";
     char *temp = NULL;
     char *context_ptr = NULL;
@@ -374,7 +374,7 @@ void flush_map(FILE *file, Map *map) {
     char line[1024];
     int rows_count = 0;        // 帧结构行数计数
     int stage_counts[9] = {0}; // 工作台各类型数量计数
-    float temp_arr[10];        // 开辟解析数据用的临时空间
+    double temp_arr[10];        // 开辟解析数据用的临时空间
     //while (fgets(line, sizeof line, stdin)) {
     while (fgets(line, sizeof line, file)) {
         if (line[0] == 'O' && line[1] == 'K') {
